backend/requirements.txt
fastapi
pydantic
uvicorn
paho-mqtt
